Move media session implementation to exoplayer module
