+++ playback/exoplayer/src/main/kotlin/session/AndroidMediaService.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.playback.core.mediasession
+package org.jellyfin.playback.exoplayer.session
 
 import androidx.media3.session.MediaSession
 import androidx.media3.session.MediaSessionService
